shadow-dom/declarative/declarative-shadow-dom-repeats.tentative.html

<!DOCTYPE html>
<title>Declarative Shadow DOM Element Attachment</title>
<link rel='author' href='mailto:masonf@chromium.org'>
<link rel='help' href='https://github.com/whatwg/dom/issues/1235'>
<link rel='help' href='https://github.com/whatwg/html/pull/10069'>
<link rel='help' href='https://github.com/whatwg/dom/pull/1246'>

<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='../../html/resources/common.js'></script>
<script src="support/helpers.js"></script>

<div id=multiple1>
  <template shadowrootmode=open>Open</template>
  <template shadowrootmode=closed>Closed</template>
</div>

<div id=multiple2>
  <template shadowrootmode=closed>Closed</template>
  <template shadowrootmode=open>Open</template>
</div>

<script>
test((t) => {
  let shadow = multiple1.shadowRoot;
  assert_true(!!shadow,'Remaining shadow root should be open');
  assert_equals(shadow.textContent,"Open");
  shadow = multiple2.shadowRoot;
  assert_false(!!shadow,'Remaining shadow root should be closed');
  multiple1.remove(); // Cleanup
  multiple2.remove();
},'Repeated declarative shadow roots keep only the first');
</script>

<div id=open1>
  <template shadowrootmode=open>Open</template>
</div>

<script>
test((t) => {
  assert_throws_dom("NotSupportedError",() => {
    open1.attachShadow({mode: "closed"});
  },'Mismatched shadow root type should throw');
  const initialShadow = open1.shadowRoot;
  const shadow = open1.attachShadow({mode: "open"}); // Shouldn't throw
  assert_equals(shadow,initialShadow,'Same shadow should be returned');
  assert_equals(shadow.textContent,'','Shadow should be empty');
},'Calling attachShadow() on declarative shadow root must match type');
</script>

<div id=open2>
  <template shadowrootmode=open shadowrootdelegatesfocus shadowrootclonable serializable>
    Open, delegates focus (not the default), clonable (not the default)
    serializable (not the default), named slot assignment (the default)
  </template>
</div>

<script>
test((t) => {
  // These should match the declarative shadow root above.
  const originalProps = {mode: "open", delegatesFocus: true, slotAssignment: "named", clonable: true, serializable: true};

  // Changing these should throw.
  assert_throws_dom("NotSupportedError",() => {
    open2.attachShadow({...originalProps, mode: "closed"});
  },'Mismatched shadow root type should throw');
  assert_throws_dom("NotSupportedError",() => {
    open2.attachShadow({...originalProps, delegatesFocus: false});
  },'Mismatched shadow root delegatesFocus should throw');
  assert_throws_dom("NotSupportedError",() => {
    open2.attachShadow({...originalProps, slotAssignment: "manual"});
  },'Mismatched shadow root slotAssignment should throw');

  // Boolean arguments just update the shadow root, and return the same root.
  const initialShadow = open2.shadowRoot;
  assert_true(open2.shadowRoot.clonable);
  assert_true(open2.shadowRoot.serializable);
  const shadow = open2.attachShadow({...originalProps, clonable: false, serializable: false}); // Shouldn't throw
  assert_equals(shadow,initialShadow,'Same shadow should be returned');
  assert_equals(shadow.textContent,'','Shadow should be empty');
  assert_false(open2.shadowRoot.clonable,'clonable flag should be updated');
  assert_false(open2.shadowRoot.serializable,'serializable flag should be updated');
},'Calling attachShadow() on declarative shadow root must match all parameters');
</script>
